Compile default files into program binary
Default files that can be included in the generated project folder are
now compiled into the haru binary.

Purpose behind this is allowing haru to remain as a single file
executable with no external dependencies.

With this method any changes to these preset files can be made inside
`assets/` and CMake will issue a rebuild of the generated .cpp files on
any changes.